requirements.txt
opencv-python==4.8.1.78
streamlit==1.28.0
Pillow==10.0.1
numpy==1.24.3